Convert to select2 type ahead component
From our initial spike, Select2 seems like a
viable alternative to TypeAhead and Chosen
(the type ahead components used for the site now)
Aside from being a viable alternative, it is
also better suited to work with Ember because
it reacts to model changes without having to
re-initialize.

We need to convert to this library when we have time.